Extract conversion of TagDetection to ROS msg.
Both Nodelet and Node should delegate to a single class for the
main logic: the class that is now AprilTagNode. This commit starts to
clean up AprilTagNode to make it easier to turn into a library. We need to
distill the class into its main task that will be required by Nodelet and
Node, and we start by extracting the tagDetectionToMsg function.